% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalMutualInformation.R
\name{cmitable}
\alias{cmitable}
\title{Conditional mutual information table}
\usage{
cmitable(x, y)
}
\arguments{
\item{x}{Input data}

\item{y}{Class vector of the same length as data}

\item{laplace}{Laplace smoothing parameter}
}
\value{
Returns table containing conditional mutial information between each attribute
}
\description{
Conditional mutual information table
}
