% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxSpanningTree.R
\name{maxSpanningTree}
\alias{maxSpanningTree}
\title{Maximum Spanning Tree}
\usage{
maxSpanningTree(table)
}
\arguments{
\item{table}{Adjacency matrix containing weights}
}
\value{
Adjacency matrix representing directed maximal spanning tree
}
\description{
Builds directed maximum (weight) spanning tree
on an undirected weigted graph described by an adjacency matrix.
Root node in resulting tree is always node number 1.
}

