% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanBayes.R
\name{predict.tanBayes}
\alias{predict.tanBayes}
\title{Predict function for TAN Bayes}
\usage{
\method{predict}{tanBayes}(object, newdata, type = c("class", "raw"),
  threshold = 0.001, eps = 0, ...)
}
\arguments{
\item{object}{object of tanBayes class}

\item{newdata}{Test data}

\item{type}{not used}

\item{threshold}{probability threshold applied to values not present in training data}

\item{eps}{minimal probability to be relplaced with treshold}
}
\description{
Predict function for TAN Bayes
}
