% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanBayes.R
\name{tanBayes.default}
\alias{tanBayes.default}
\title{TAN Bayes Classifier}
\usage{
\method{tanBayes}{default}(x, y, laplace = 0, ...)
}
\arguments{
\item{x}{Input data}

\item{y}{Class vector}

\item{laplace}{Laplace smoothing parameter}

\item{...}{Not used}
}
\value{
tanBayes object
}
\description{
TAN Bayes Classifier
}
