% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanBayes.R
\name{tanBayes.formula}
\alias{tanBayes.formula}
\title{TAN Bayes Classifier}
\usage{
\method{tanBayes}{formula}(formula, data, laplace = 0, ..., subset,
  na.action = na.pass)
}
\arguments{
\item{formula}{Formula object}

\item{data}{Input data}

\item{laplace}{Laplace smoothing parameter}

\item{...}{not used}
}
\value{
tanBayes object
}
\description{
TAN Bayes Classifier
}
