% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalMutualInformation.R
\name{cmitable}
\alias{cmitable}
\title{Conditional mutual information}
\usage{
cmitable(x, y)
}
\arguments{
\item{x}{Input data}

\item{y}{Class vector of the same length as data}
}
\value{
Returns table containing conditional mutual information
   calculated for each pair of attributes
}
\description{
Calculates conditional mutual information
for each pair of dataset attributes
}

