% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxSpanningTree.R
\name{maxSpanningTree}
\alias{maxSpanningTree}
\title{Returns directed maximal spanning tree
as adjacency matrix}
\usage{
maxSpanningTree(table)
}
\arguments{
\item{table}{Adjacency matrix containing weights}
}
\value{
Adjacency matrix representing directed maximal spanning tree
}
\description{
Returns directed maximal spanning tree
as adjacency matrix
}
