% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanBayes.R
\name{tanBayes}
\alias{tanBayes}
\title{Tree Augmented Naive (TAN) Bayes classifier}
\usage{
tanBayes(data)
}
\description{
Tree Augmented Naive (TAN) Bayes classifier
}
\examples{
tanBayes(data)
}
\keyword{Bayes}
\keyword{TAN}
